Update configs to catch up with recent website updates (#27)
* Fix tests not running with test profile

The PHPUnit config was moved to its default location: the project's
root. As such, it no longer has to be passed in as an argument.

* Update commerce in included config

The shape of and data in the commerce config was updated as part of an
updated that introduced the Black Friday sale.

* Fix database encoding errors

Connecting to the database with the `utf8mb4` character set offers
better emoji support.